move parameters & update_requirements to EvoOptimiser
